[OptReport] Introduce mechanism to represent unique remark ID in LoopOptRemark.
Every remark emitted in a LoopOptReport is expected to have a
unique diagnostic ID which will be used while cataloging various
remark messages. Currently LoopOptReport framework ignores these
IDs and emits LoopOptRemark in below format -
!1 = !{!"intel.optreport.remark", !<Remark Msg>, <Remark Args>}

This patch adds new interfaces in LoopOptReport to explicitly
add a remark's unique ID to LoopOptRemark metadata. The current
interfaces are indicated as legacy code and all users of
LoopOptReport are expected to switch to the new interfaces. VPlan
vectorizer has been updated accordingly as an example. The new
format for LoopOptRemark is -
!1 = !{!"intel.optreport.remark", !<Remark ID>, !<Remark Msg>, <Remark Args>}

Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/305367
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>

+++ llvm/lib/Transforms/Intel_LoopTransforms/HIRLoopDistribution.cpp
@@ -1,6 +1,6 @@
 //===----- HIRLoopDistribution.cpp - Distribution of HIR loops  -----------===//
 //
-// Copyright (C) 2015-2020 Intel Corporation. All rights reserved.
+// Copyright (C) 2015-2021 Intel Corporation. All rights reserved.
 //
 // The information and source code contained herein is the exclusive
 // property of Intel Corporation and may not be disclosed, examined
@@ -830,7 +830,8 @@ void HIRLoopDistribution::distributeLoop(
       }
     }
 
-    LORBuilder(*LoopNode).addOrigin("Distributed chunk %d", List.index() + 1);
+    LORBuilder(*LoopNode).addOrigin("Distributed chunk %d",
+                                    (int)List.index() + 1);
   }
 
   // The loop is now empty, all its children are moved to new loops
